refactor: /ui/use/ → /ui.use/
