+++ code/ext/ext.lib.automerge/src/index.mts
@@ -38,7 +38,7 @@ export { HistoryGrid } from './ui/ui.History.Grid';
 export { Info, InfoField } from './ui/ui.Info';
 export { RepoList } from './ui/ui.RepoList';
 export { CmdBar } from './ui/ui.Cmd.Bar';
-export { Redraw, useDoc, useDocs, useRedrawOnChange } from './ui/use';
+export { Redraw, useDoc, useDocs, useRedrawOnChange } from './ui.use';
 
 /**
  * Dev
