+++ code/ext/ext.lib.automerge/src/types.ts
@@ -22,7 +22,7 @@ export type * from './ui/ui.RepoList.Model/t';
 export type * from './ui/ui.RepoList.Virtual/t';
 export type * from './ui/ui.RepoList/t';
 
-export type * from './ui/use/t';
+export type * from './ui.use/t';
 
 /**
  * Automerge JS object extensions.
