+++ code/ext/ext.lib.automerge/src/ui.use/common.ts
@@ -1,0 +1,2 @@
+export * from '../ui/common';
+export { Doc } from '../crdt';
