+++ code/ext/ext.lib.automerge/src/ui/ui.Info/ui.History.tsx
@@ -1,4 +1,4 @@
-import { useHistory } from '../use';
+import { useHistory } from '../../ui.use';
 import { Value, css, type t } from './common';
 
 export type HistoryProps = {
