+++ code/ext/ext.lib.automerge/src/ui/ui.Info/use.Data.ts
@@ -1,4 +1,4 @@
-import { useDocs } from '../use';
+import { useDocs } from '../../ui.use';
 import { Is, type t } from './common';
 import { Data } from './u';
