--- code/ext/ext.lib.automerge/src/ui/use/common.ts
@@ -1,1 +1,0 @@
-export * from '../common';
